Fixed paging and added extra community wrapper.
There was a bug with GetPagesByTotal() where the total amoutn of pages
would not be calculated right due to a casting error. This casting error
has been fixed and the method should now request the right numebr of
pages. Also, GetCommunityStreams() was added which gets all streams
currently streaming in a community. This extra wrapper is found under
the 'Streams' endpoint.