Started implementing Exceptions and a few more minor bug fixes.
I started to implement exceptions when creating an instance of
TwitchIrcClient for formatting and validity checks. This will soon be
implemented library wide as things are more concrete.

Fixed a formatting error with Send() that prevented messages from
actually being sent.

Removed PrivmsgTemplate and WhisperTemplate and replaced them with
MessageTemplate seeing as they were identical structures.